==> 01_MSSD_calculation/03_Combing_EA_task_regional_MSSD.ipynb <==
import glob
import pandas as pd
import os
import scipy.stats as stats
import numpy as np
import matplotlib.pyplot as plt

'''MERGING SPINS AND SPASDS DATASET CSVS'''

df1 = pd.read_csv('/projects/tsecara/SPINS_ASD_Project2/data/SPASD/mssd_meancentered_SPASD_EA_RSclean.csv')
df2 = pd.read_csv('/projects/tsecara/SPINS_ASD_Project2/data/SPINS/mssd_meancentered_SPINS_EA_RSclean.csv') 

# Concatenate DataFrames vertically
merged_df = pd.concat([df1, df2], ignore_index=True)

# Delete a specific row by index
rows_to_drop = [123]  # deletes the matrix region label name
merged_df = merged_df.drop(rows_to_drop)

# Delete the first column
merged_df = merged_df.drop(merged_df.columns[0], axis=1)
# Save the merged DataFrame to a new CSV file
merged_df.to_csv('/projects/tsecara/SPINS_ASD_Project2/data/combined/combined_mssd_mean_EA_RSclean.csv', index=False)
